Refactored the REPL to use the new ose_bindcfn function to bind its functions into the environment, replacing the old serve() function.
Important: This commit contains a change to osevm_run(), which now
checks the number of elements in the dump, and only restores them
if they were added after osevm_run was entered. There are scenarios
where we want to store the current state of the stacks in the dump,
call run more than once, and only then restore the state. For example:
in the REPL, if the user is stepping through computation and wants
to enter something, the lexer and parser need to call osevm_run
more than once, and if that function aggressively restores the dump
only exiting once it is empty, the state will be restored too soon.

osevm_step() was not altered since it implies some machinery around it
checking the conditions to continue.