JS transpilation support.
Most of ose_stackops.c has been wrapped in ose_wrapper.js, and
all functions are exported via the list in osejs_export.mk.

As of this commit, when the osejs code is loaded, the VM is
created and loaded as well. This will likely be separated
in a future commit so that the library can be used simply as an
OSC implementation in the browser without VM suppor if it's
not needed.