Removed the lexer and parser.
The lexer and parser made it possible to specify a number of addresses
all in a single string, rather than entering them separately, line
by line. This was confusing for a number of reasons, and really
wasn't more than a weak ad hoc language, so it's been taken out.
Now and for the time being, the only way to enter a sequency
of addresses is to "send" them one at a time, which is the
cannonical way.