A number of changes to support /!/eval and /!/apply.
-- /!/eval now can take an optional second argument that will serve as the
environment
-- /!/eval leaves a copy of the environment on top of the stack when it
finishes
-- Fixed a control bug: when moving everything to the dump in preparation
to eval or apply, first, control wasn't being moved, which was a bug, but
when that behavior was changed, that left control empty causing the vm to
call /!/drop on an empty bundle and trigger an assertion. The new behavior
is that the VM now simply checks to see if there's anything in control
before calling /!/drop, which I believe is the right thing to do. Other options
include stuffing a dummy there before eval or apply return, but that's
some weird kludgy behavior for other functions to have to do should they
implement this kind of behavior. /!/drop could be made permissive, but I'd
prefer to keep it as is now, since most calls to /!/drop on an empty bundle
are probably unintentional.
-- Fixed a bug in the REPL that was causing the history file to never be
written. There's now a longjmp out of the signal handler which cleans up
before exiting.
-- Added a first version of /> which will no doubt change in a later commit.
-- Removed (commented out) the second (env) UDP port from the REPL.