Reorganized the conf system to better support Arduino.
All actual conf parameters are set in ose.h which must be included
by everything. ose.h includes ose_conf.h which doesn't exist when the
repo is cloned. There's a file called ose_conf.h.example that can either
be renamed, or will be copied to ose_conf.h when the Makefile
is run.