Updates to Teensy sketch with supporting changes to core files.
Teensy sketch currently uses Serial1 to communicate (Serial to
be added back in a future commit).

Modified the SLIP code in ose_util.c to also produce an empty
message when a cr or nl is received before a null byte has been received.

Changed the name of ose_inputTopLevel to ose_inputMessages, and
added a similar ose_inputMessage for a single message that's not
in a bundle.

Added \r to the end of strings in ose_print.c.